Merge develop to master

+++ mat-server-commons/src/main/java/mat/server/util/MeasureUtility.java
@@ -1,6 +1,3 @@
-/**
- * 
- */
 package mat.server.util;
 
 import mat.shared.StringUtility;
@@ -129,15 +126,15 @@ public class MeasureUtility {
 		String cleanedString = "";	
 		for(int i=0;i<originalString.length();i++){
 			char c = originalString.charAt(i);
-			int intc = (int)c;
-			
-			if(c == '_' || (intc >= 48 && intc <= 57) || (intc >= 65 && intc <= 90) || (intc >= 97 && intc <= 122)){
-				
-				if(!(cleanedString.isEmpty() && Character.isDigit(c))){
+			int intc = c;
+
+			if (c == '_' || (intc >= 48 && intc <= 57) || (intc >= 65 && intc <= 90) || (intc >= 97 && intc <= 122)) {
+
+				if (!(cleanedString.isEmpty() && Character.isDigit(c))) {
 					cleanedString = cleanedString + "" + c;
 				}
-				
-			} 
+
+			}
 
 		}
 		
@@ -145,7 +142,7 @@ public class MeasureUtility {
 	}
 	
 	/**
-	 * Method to set latest QDM Version in Draft's or clones of CQL type measure or CQL Stand Alone Library.
+	 * Method to set latest QDM Version in Draft's or clones of CQL type measure or CQL Measure.
 	 * 
 	 * **/
 	public static void updateLatestQDMVersion(XmlProcessor processor) throws XPathExpressionException {
